refactor: change directory name from model, entity to domain
