+++ src/main/java/bitnagil/bitnagil_backend/auth/kakao/service/CustomOAuth2UserService.java
@@ -11,11 +11,11 @@ import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 
-import bitnagil.bitnagil_backend.auth.kakao.model.CustomOAuth2User;
-import bitnagil.bitnagil_backend.auth.kakao.model.OAuth2Attribute;
+import bitnagil.bitnagil_backend.auth.kakao.domain.CustomOAuth2User;
+import bitnagil.bitnagil_backend.auth.kakao.domain.OAuth2Attribute;
 import bitnagil.bitnagil_backend.user.Repository.UserRepository;
 import bitnagil.bitnagil_backend.enums.SocialType;
-import bitnagil.bitnagil_backend.user.entity.User;
+import bitnagil.bitnagil_backend.user.domain.User;
 import lombok.RequiredArgsConstructor;
 
 /**
