+++ src/main/java/bitnagil/bitnagil_backend/user/service/UserAuthService.java
@@ -13,7 +13,7 @@ import bitnagil.bitnagil_backend.global.exception.CustomException;
 import bitnagil.bitnagil_backend.user.Repository.UserRepository;
 import bitnagil.bitnagil_backend.enums.SocialType;
 import bitnagil.bitnagil_backend.auth.jwt.TokenResponse;
-import bitnagil.bitnagil_backend.user.entity.User;
+import bitnagil.bitnagil_backend.user.domain.User;
 import bitnagil.bitnagil_backend.enums.Role;
 import bitnagil.bitnagil_backend.auth.kakao.response.KakaoAccount;
 import bitnagil.bitnagil_backend.auth.kakao.response.KakaoUserInfoResponse;
